Initial Commit
Creating a simple Kofi donor thank you scroll tool for consuming the CSV that KoFi exports and generating a scroll for it. This will come in handy when I make videos. Hoping to sell compiled version of this project as well, need to add GNU-GPL3 license to this. Compiling it yourself and you can have it for free, I'm just counting on people not wanting to put in the effort